+++ src/Sedona/SBORuntimeBundle/DependencyInjection/SedonaSBORuntimeExtension.php
@@ -11,17 +11,18 @@
 
 namespace Sedona\SBORuntimeBundle\DependencyInjection;
 
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 /**
  * This is the class that loads and manages your bundle configuration.
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
  */
-class SedonaSBORuntimeExtension extends Extension
+class SedonaSBORuntimeExtension extends Extension implements PrependExtensionInterface
 {
     /**
      * {@inheritdoc}
@@ -33,5 +34,26 @@ class SedonaSBORuntimeExtension extends Extension
 
         $loader = new Loader\XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.xml');
+    }
+
+    /**
+     * Allow an extension to prepend the extension configurations.
+     *
+     * @param ContainerBuilder $container
+     */
+    public function prepend(ContainerBuilder $container)
+    {
+        $bundles = $container->getParameter('kernel.bundles');
+
+        if (isset($bundles['TwigBundle'])) {
+            $container->prependExtensionConfig(
+                'twig',
+                array(
+                    'form_theme'    => array(
+                        'SedonaSBORuntimeBundle:layout:sbo-form-theme.html.twig'
+                    )
+                )
+            );
+        }
     }
 }